complete get surroundings with calcutaion on server (need to switch to DB) complete get rooms by type id complete get groups by vendor Id

+++ src/main/java/org/avengers/capstone/hostelrenting/dto/HostelGroupDTO.java
@@ -19,9 +19,9 @@ public class HostelGroupDTO implements Serializable {
 
     private String street;
 
-    private Float longitude;
+    private Double longitude;
 
-    private Float latitude;
+    private Double latitude;
 
     /**
      * curfewTime that limit the time to come back home
